moved setM5 to testUtils so that sims_catUtils unit tests can import it
